drop _urls module
